Fixed width the modal window
Fixed the base value of the "max-width" property in the "modal-window" class from 450px to 528px.
With the base value the "max-width" property of 450px in the "modal-window" class, the value of 528px the "width" property in media query (min-width: 1200px) was not applied. Accordingly, with a viewport width of 1200px, the width of modal windoww was 450px. A contant protruded out of bounds the modal window.